backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
openai==1.99.9
python-dotenv==1.1.1
supabase==2.15.3
